Add files via upload
SimpleUI.py opens an UI that can be started to predict a temperature and stopped at any time. The data needed for the model to predict the temperature is gathered via mqtt and isnt included here. Sample data, that shows how the dataframe looks like, can be found in the test function inside sensor_mqtt.py